+++ src/NetscriptFunctions.ts
@@ -24,6 +24,7 @@ import {
   LocationName,
   ToastVariant,
   UniversityClassType,
+  CompanyName,
 } from "@enums";
 import { PromptEvent } from "./ui/React/PromptManager";
 import { GetServer, DeleteServer, AddToAllServers, createUniqueRandomIp } from "./Server/AllServers";
@@ -86,7 +87,7 @@ import { Flags } from "./NetscriptFunctions/Flags";
 import { calculateIntelligenceBonus } from "./PersonObjects/formulas/intelligence";
 import { CalculateShareMult, StartSharing } from "./NetworkShare/Share";
 import { recentScripts } from "./Netscript/RecentScripts";
-import { InternalAPI, removedFunction, NSProxy } from "./Netscript/APIWrapper";
+import { InternalAPI, setRemovedFunctions, NSProxy } from "./Netscript/APIWrapper";
 import { INetscriptExtra } from "./NetscriptFunctions/Extra";
 import { ScriptDeath } from "./Netscript/ScriptDeath";
 import { getBitNodeMultipliers } from "./BitNode/BitNode";
@@ -111,6 +112,7 @@ export const enums: NSEnums = {
   LocationName,
   ToastVariant,
   UniversityClassType,
+  CompanyName,
 };
 for (const val of Object.values(enums)) Object.freeze(val);
 Object.freeze(enums);
@@ -824,10 +826,7 @@ export const ns: InternalAPI<NSFull> = {
           ++scriptsKilled;
         }
       }
-      helpers.log(
-        ctx,
-        () => `Killing all scripts on '${server.hostname}'. May take a few minutes for the scripts to die.`,
-      );
+      helpers.log(ctx, () => `Killing all scripts on '${server.hostname}'.`);
 
       return scriptsKilled > 0;
     },
@@ -1162,7 +1161,11 @@ export const ns: InternalAPI<NSFull> = {
 
     const cost = getPurchaseServerCost(ram);
     if (cost === Infinity) {
-      helpers.log(ctx, () => `Invalid argument: ram='${ram}'`);
+      if (ram > getPurchaseServerMaxRam()) {
+        helpers.log(ctx, () => `Invalid argument: ram='${ram}' must not be greater than getPurchaseServerMaxRam`);
+      } else {
+        helpers.log(ctx, () => `Invalid argument: ram='${ram}' must be a positive power of 2`);
+      }
       return Infinity;
     }
 
@@ -1767,7 +1770,7 @@ export const ns: InternalAPI<NSFull> = {
   }),
   getFunctionRamCost: (ctx) => (_name) => {
     const name = helpers.string(ctx, "name", _name);
-    return getRamCost(...name.split("."));
+    return getRamCost(name.split("."), true);
   },
   tprintRaw: () => (value) => {
     Terminal.printRaw(wrapUserNode(value));
@@ -1778,9 +1781,10 @@ export const ns: InternalAPI<NSFull> = {
   flags: Flags,
   ...NetscriptExtra(),
 };
-// Object.assign to bypass ts for removedFunctions which have no documentation or ramcost
-Object.assign(ns, {
-  getServerRam: removedFunction("v2.2.0", "getServerMaxRam and getServerUsedRam"),
+
+// Removed functions
+setRemovedFunctions(ns, {
+  getServerRam: { version: "2.2.0", replacement: "getServerMaxRam and getServerUsedRam" },
 });
 
 export function NetscriptFunctions(ws: WorkerScript): NSFull {